requirements.txt
python-telegram-bot==21.1.1
yt-dlp>=2024.07.25
requests==2.32.3